backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
mysql-connector-python==8.4.0
selenium==4.22.0
webdriver-manager==4.0.2
